ibook_backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
boto3==1.34.39
botocore==1.34.39
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
jmespath==1.0.1
multidict==6.0.5
mysqlclient==2.2.1
openai==0.28.0
pillow==10.2.0
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
yarl==1.9.4
